p1/video/requirements.txt
torch==2.2.2
timm>=0.9.16,<2
transformers==4.38.2
huggingface_hub==0.21.4
opencv-python-headless>=4.9.0.80
Pillow>=10.2.0
numpy>=1.23.5,<2
